Subscriptions: Fix regression to not subscribe with no app
Due to the changes to the subscription machinery to support
HotReloading, a regression was introduced whereby declarative
subscriptions would be started, even if the app was no present or
healthy, leading to attempts to process pubsub messages and potentially
panic.

Updates subscription machinery to not start declarative subscriptions if
the app is not present or healthy.

Adds integration tests to cover these cases.

Signed-off-by: joshvanl <me@joshvanl.dev>