Fix: Don't track duplicate in-flight reminders for non-scheduler reminders (#7998)
* Fix: Don't track duplicate in-flight reminders for non-scheduler
reminders

Don't track in-flight internal (workflow) reminders when not using
scheduler reminders. This is because they only get un-tracked when they
are deleted in the scheduler code-path. This caused workflow activities
to not fire when triggering a workflow multiple times with the same
instance ID - since the had the same reminder key.

Adds int tests for both scheduler and non-scheduler reminders which
trigger a workflow with the same instance ID multiple times.

Signed-off-by: joshvanl <me@joshvanl.dev>

* Wrap state check in Eventually

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>
Co-authored-by: Artur Souza <asouza.pro@gmail.com>