Scheduler: Adds Kubernetes namespace controller
When in Kubernetes mode, the Scheduler will delete Jobs of all kinds
(currently cron and actor reminders) when the namespace they belong to
is deleted. This prevents both squatting memory in Etcd, and more importantly,
consuming resources scheduling jobs which have no consumers to signal.

The Scheduler Namespace Controller will Delete all Jobs in a namespace
when the Kubernetes Namespace is deleted. The controller uses leader election,
meaning only one instance of the Scheduler will be running the controller at
any one time.

Uses an updated go-etcd-cron library version to use the new
`DeletePrefixes` API method.

The Scheduler has been given the permissions to get, list and watch
namespaces in Kubernetes so the controller can function. The controller
is only executed when the Scheduler is in Kubernetes mode
(`--mode=kubernetes`).

A new `--kubeconfig` CLI flag has been added to the Scheduler, which is
used in integration testing so that it points to the mock Kubernetes
API server. This isn't useful for end users so hasn't been documented or
exposed in the Helm chart.

Since the Namespace controller needs to call the Etcd Cron library, the
Scheduler server has been refactored to move functionality into
`/internal` of the server package, so that they can be shared between
both the server and controller.

Signed-off-by: joshvanl <me@joshvanl.dev>