Remove internal actors
Begone spaghetti monster.

This PR refactors the actor runtime with the purpose of facilitating
proper durable dynamic workflow actor support. The initial intention was
to remove the concept of "internal actors" as this abstraction is
wrong for an actor engine, and instead replace it with "actor targets"
(app, workflow, activity, and scope for extensions). This removes any
magic codes paths and creates proper abstractions. Due to the code state
actors was in, this change cascaded into a refactor of actors into
multiple packages which combine into the full actor runtime. The current
actor runtime should now be more extensible, maintainable, and support
features required for workflows, like dynamic actor type registration.

- Actor Placement has been overhauled to now correctly lock the entire
  actor runtime during dissemination events.
- Reminders and timers have been split out into interfaces
- Actor targets have been introduced, which are the target
  implementation for app/workflow.
- Table tracks the registered actor types and current active actors.
- Engine is the main entry point for actors, and is responsible for
  managing the execution of actors.
- API entries into actors now hang until the actor runtime is ready.
- Actor reentry uses a fifo circle buffer.

Workflows continue to be in a non-fully functional state, and are
infact, likely more broken due to these changes. This is expected and is
going to be addressed in a follow change now that we have a good actor
runtime foundation.

A large set of integration tests have been added for actor
functionality.

Signed-off-by: joshvanl <me@joshvanl.dev>