Workflows: Dynamic Actor Types & No Component
Branched from https://github.com/dapr/dapr/pull/8298

This change focuses on 2 things; making Workflow actor type registration
dynamic and the removal of the Workflow backend and internal Workflow
component.

To ensure that hosts only advertise that they are able to complete
workflow work, the workflow actor types are registered when the first
workflow receive work items client has connected, and un-register the
workflow actor types once there are no process workflow items connected.
This means no actor reminders will be scheduled to that node when there
are no workflow items connected.

The Workflow components have been removed. The sole reason this existed
was to facilitate an experimentation for using sqlite as the durable
message layer for workflow backends. This mode is orthogonal with the
direction of the project at the moment as it does not give the same
distributed async non-localized properties that Scheduler provides. For
this reason, sqlite has been removed, as well as all the downstream code
it required. This code added a lot of magic code complexity and
maintenance. Sqlite as a supported backend was neither documented, nor
even shipped in a released build of Dapr. It and all mentions have been
removed.

Signed-off-by: joshvanl <me@joshvanl.dev>