Workflows: Dynamic Actor Types & No Component (#8310)
* Remove internal actors

Begone spaghetti monster.

This PR refactors the actor runtime with the purpose of facilitating
proper durable dynamic workflow actor support. The initial intention was
to remove the concept of "internal actors" as this abstraction is
wrong for an actor engine, and instead replace it with "actor targets"
(app, workflow, activity, and scope for extensions). This removes any
magic codes paths and creates proper abstractions. Due to the code state
actors was in, this change cascaded into a refactor of actors into
multiple packages which combine into the full actor runtime. The current
actor runtime should now be more extensible, maintainable, and support
features required for workflows, like dynamic actor type registration.

- Actor Placement has been overhauled to now correctly lock the entire
  actor runtime during dissemination events.
- Reminders and timers have been split out into interfaces
- Actor targets have been introduced, which are the target
  implementation for app/workflow.
- Table tracks the registered actor types and current active actors.
- Engine is the main entry point for actors, and is responsible for
  managing the execution of actors.
- API entries into actors now hang until the actor runtime is ready.
- Actor reentry uses a fifo circle buffer.

Workflows continue to be in a non-fully functional state, and are
infact, likely more broken due to these changes. This is expected and is
going to be addressed in a follow change now that we have a good actor
runtime foundation.

A large set of integration tests have been added for actor
functionality.

Signed-off-by: joshvanl <me@joshvanl.dev>

* go mod tidy

Signed-off-by: joshvanl <me@joshvanl.dev>

* Adds review comments

Signed-off-by: joshvanl <me@joshvanl.dev>

* Workflows: Dynamic Actor Types & No Component

Branched from https://github.com/dapr/dapr/pull/8298

This change focuses on 2 things; making Workflow actor type registration
dynamic and the removal of the Workflow backend and internal Workflow
component.

To ensure that hosts only advertise that they are able to complete
workflow work, the workflow actor types are registered when the first
workflow receive work items client has connected, and un-register the
workflow actor types once there are no process workflow items connected.
This means no actor reminders will be scheduled to that node when there
are no workflow items connected.

The Workflow components have been removed. The sole reason this existed
was to facilitate an experimentation for using sqlite as the durable
message layer for workflow backends. This mode is orthogonal with the
direction of the project at the moment as it does not give the same
distributed async non-localized properties that Scheduler provides. For
this reason, sqlite has been removed, as well as all the downstream code
it required. This code added a lot of magic code complexity and
maintenance. Sqlite as a supported backend was neither documented, nor
even shipped in a released build of Dapr. It and all mentions have been
removed.

Signed-off-by: joshvanl <me@joshvanl.dev>

* Adds replace for durable task to joshvanl for callback

Signed-off-by: joshvanl <me@joshvanl.dev>

* Updates durabletask-go to dapr repo HEAD

Signed-off-by: joshvanl <me@joshvanl.dev>

* Make workflow reminders oneshot with Scheduler. Enable missing workflow
int tests

Signed-off-by: joshvanl <me@joshvanl.dev>

* Wire up scheduler reminders

Signed-off-by: joshvanl <me@joshvanl.dev>

* Re-enable noappentities healthz check

Signed-off-by: joshvanl <me@joshvanl.dev>

* Removes workflowcomponent e2e&perf tests

Signed-off-by: joshvanl <me@joshvanl.dev>

* Adds back resiliency to actor API

Signed-off-by: joshvanl <me@joshvanl.dev>

* Prevent panic in actor healthcheck

Signed-off-by: joshvanl <me@joshvanl.dev>

* Adds review comments

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix actor response error header

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix actor resiliency

Signed-off-by: joshvanl <me@joshvanl.dev>

* Skip non-compatable int test on windows

Signed-off-by: joshvanl <me@joshvanl.dev>

* Skip incompatible deactivation int test on windows

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix reentry return HTTP status code

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix actor reminder statestore lookup fn

Signed-off-by: joshvanl <me@joshvanl.dev>

* Revert omit healthcheck config

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix nil pointer

Signed-off-by: joshvanl <me@joshvanl.dev>

* Make max depth error permanent

Signed-off-by: joshvanl <me@joshvanl.dev>

* Permanent backoff reentry errors

Signed-off-by: joshvanl <me@joshvanl.dev>

* Adds scheduler reminder e2e test and fix portforward

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix actor engine backoff on 404

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix actor method not found retry

Signed-off-by: joshvanl <me@joshvanl.dev>

* Review comments

Signed-off-by: joshvanl <me@joshvanl.dev>

* Review comments

Signed-off-by: joshvanl <me@joshvanl.dev>

* Adds period actor reminder int test

Signed-off-by: joshvanl <me@joshvanl.dev>

* tmp disable actor reminder e2e test for debug

Signed-off-by: joshvanl <me@joshvanl.dev>

* Add wait groups to reminder execution stop start

Signed-off-by: joshvanl <me@joshvanl.dev>

* Remove wait for onplacementupdate

Signed-off-by: joshvanl <me@joshvanl.dev>

* Make reminder drain rebalance all

Signed-off-by: joshvanl <me@joshvanl.dev>

* Always tick reminder when called

Signed-off-by: joshvanl <me@joshvanl.dev>

* Add review comments

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix backoff import of actor reminder e2e test

Signed-off-by: joshvanl <me@joshvanl.dev>

* Ensure reminder loop is stoppped before starting again

Signed-off-by: joshvanl <me@joshvanl.dev>

* Make actor features use scheduler

Signed-off-by: joshvanl <me@joshvanl.dev>

* Ensure reminder maps are cleared on drain

Signed-off-by: joshvanl <me@joshvanl.dev>

* Remove from activeremidners in loop

Signed-off-by: joshvanl <me@joshvanl.dev>

* Ensure reminder ticks always

Signed-off-by: joshvanl <me@joshvanl.dev>

* Revert statestore code

Signed-off-by: joshvanl <me@joshvanl.dev>

* Revert scheduler

Signed-off-by: joshvanl <me@joshvanl.dev>

* Increase period after scheduler lag on CRUD events

Signed-off-by: joshvanl <me@joshvanl.dev>

* Make actor e2e use scheduler

Signed-off-by: joshvanl <me@joshvanl.dev>

* Adds output of actor log

Signed-off-by: joshvanl <me@joshvanl.dev>

* .

Signed-off-by: joshvanl <me@joshvanl.dev>

* .

Signed-off-by: joshvanl <me@joshvanl.dev>

* .

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>