sentry: allow for self-hosted CA on Kubernetes
Currently, any user running on kubernetes interesting on running
a standalone Sentry is forced to provide both a `dapr-trust-bundle`
Secret and ConfigMap. This is not ideal in the cases where you're
generating the private key pair from eg. cert-manager and distributing
trust in the cluster using trust-manager. On these use-cases it's simply
enough to mount these resources and supply the correct CLI args.

Signed-off-by: Luis Rascao <luis.rascao@gmail.com>