Workflow: gracefully recover from legacy state read errors (#8364)
Currently, after upgrading from Dapr v1.14 to v1.15 and executing a
workflow which has state from v1.14, the workflow state will fail to
parse and cause the workflow to error. This is due to the storage API
changing moving to v1 GA.

To resolve this, we fall back to parsing the workflow metadata key as
JSON, then in the case of event history, ignoring the existing state if
it fails to parse. Practically, this means a mid-running workflow will
be abandoned when upgrading to v1.15 which seems acceptable. The other
approach to this would be for use intervention to manually clean up the
actor state store which is unreasonable.

Signed-off-by: joshvanl <me@joshvanl.dev>