Workflow: Stream runtime status to durabletask, rather than poll (#8369)
* Workflow: Stream runtime status to durabletask, rather than poll

Instead of constituently polling for the runtime status to change state
to observe a different phase of a workflow, durabletask now streams the
state, whereby updates are only sent when they occur. Updates dapr to
support this.

A new "generic" `CallActorStream` internal Dapr proto has been added
which implements a internal invoke request stream that receives invoke
responses. This RPC is used to stream the runtime status of a workflow
which is a hosted on another dapr host, i.e. a remote running workflow.

Creates `CallStream` will child local and remote calls to the target
actor to implement the new remote RPC, as well as the entry point for
the workflow actor backend implementation. Only the "workflow" actor
target implements this RPC to watch for runtime state changes.

Upon invoke stream invocation, the workflow actor target will first lock
the actor to ensure we get the current orchestration metadata state,
then adds a subscription to metadata changes, sends the current state,
then unlocks so that the workflow can continue to execute. Upon changes,
the client will receive changes.

Relies on kit and durabletask PR changes.

Signed-off-by: joshvanl <me@joshvanl.dev>

* Set deps to dapr upstream

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>
Co-authored-by: Yaron Schneider <schneider.yaron@live.com>