Workflow: Don't locak placement for Activity Reminders
Activity reminders are executed to execute an actual activity. These are
often running for some non-trival period of time, and are sensitive to
double executions on the app. For this reason, we want to avoid
cancelling them as much as possible.

Because Scheduler push reminders to apps and they ensure they are only
executed one at a time, we can rely on this behaviour that they will
only execute on one app at a time. This removes the need of locking
placement to ensure single execution.

When placement dissemination happens, which is often during the start-up
of multiple replicas running workflows, we result in many double
execution of activities.

Update the runtime scheduler streamer so that reminders for
Activities will skip placement lock, preventing unnecessary double
executions.

Signed-off-by: joshvanl <me@joshvanl.dev>