Workflows: CRUD API sync
Due to the update of remote push based Scheduler for implementing
workflow commands, the existing (non-durabletask gRPC client) crud
operations will not have the command actually execute before the call
returns to the client.

Update the workflow universal APIs to wait for the desired state change
before returning the call to the client.

Fixes e2e tests which expect universal API calls to be sync.

Signed-off-by: joshvanl <me@joshvanl.dev>