Actors: Move reentrancy lock to top level
Moves the reentrnacy actor lock to the top level of the API call stack
in the engine. This is done so that in the event of a placement
dissemination during 2 calls of the same actor ID, after unlocking the
order of calls is preserved and there is not a non-deterministic race to
when either call is processed.

Introduces a `Locker` which is responsible for indexing actor ID locks.
Workflow activity actor reminders will continue to skip locks for
reminders when Scheduler push based reminders are used.

Fixes actor lock e2e test.

Signed-off-by: joshvanl <me@joshvanl.dev>