Scheduler: Update/Remove member hosts
When in Kubernetes mode, Scheduler will update the existing member on
boot, and remove it on shutdown. This is to attempt to prevent member ID
collisions on restarts and upgrades.

Update the client listen address to be only over localhost, removing the
need for mTLS.

Updates dapr/kit to v0.15.0

Signed-off-by: joshvanl <me@joshvanl.dev>