Actors: Add Placement client reconnect interval
Adds a 1s duration interval before attempting to reconnect to placement
in the event of a disconnect, likely due to leadership changes. This
prevents large log spam due to a tight loop in the reconnect logic,
where it will likely take some time for the new placement leader to be
elected.

Signed-off-by: joshvanl <me@joshvanl.dev>