Actors: Fix Placement client reconnect
Fixes an issue whereby a deadlock would occur when halting all actors,
which happens during shutdown or placement dissemination events. This
would happen more often during workflow execution as workflow and
activity actors have a short deactivation time, and the bug involved a
race condition of ranging over the actor table and deleting elements
from the table.

Fixes the actor placement client reconnect logic to be more robust by
cancelling a placement connection when either a `Send` or `Recv` fails.
Also dereferences the placement gRPC connection in the event of
connection refused errors to ensure that the round robin load balance
state is reset so new placement IPs can be discovered from the headless
service. These errors occur when 1 or more placement pods are completely
deleted.

Suppresses an EOF error when streaming workflow runtime status, as this
is a normal condition when the workflow actor stream has been
terminated.

Signed-off-by: joshvanl <me@joshvanl.dev>