Fix Actor & Scheduler deadlocks and Integration Test fixes
This changes includes a number of changes to prevent dead locks in the
actor runtime & Scheduler connectivity/job triggering sequences. There a
number of disparate scenarios covered below. This change ensures that no
deadlocks occur, particularly when running workflows, during scale (up
or down) events, loosing connection or destruction of the placement
and scheduler control plane services. Makes integration tests more
stable, particularly in CI/slow machine runners.

Scenarios and fixes:

Adds a new `scheduler` fields to the gRPC and HTTP metadata output,
which includes an object containing the list of addresses which the
daprd is currently connected to. This address list can change over time,
as daprd dynamically connects/looses connection to scheduler hosts. Int
tests added for this field, as well as being used to check the address
list in some integration tests where the number of hosts connected is a
dependency of that test scenario (e.g. job load balancing).

Moves the placement lock to `dapr/kit/concurrency/lock` as a new shared
`OuterCancel` lib. This struct was also fixed so that a deadlock could
not occur on `RLock` whereby the given context was cancelled, but
`RLock` did not return with error- now it does. This caused deadlocks in
workflow actors whereby dissemination (deactivation) of actors would
not complete as their was a circular lock which did not break.

Removes a deadlock in the Scheduler server whereby a new daprd stream
was stuck adding to the connection pool. This was due to adding the new
connections as deliverable prefixes, locks the underlying go-etcd-cron
queue for those prefixes. If a job trigger was occurring at the same
time, there was a circular lock with no break since the same Scheduler
lock adding the new connection was RLocked to get the connection for
that job trigger to send the job event to daprd. The use of the new
`Context` lock from `dapr/concurrency/lock` ensures that getting the
connection for a triggered job will fail the RLock if it does not gain
the lock after 250ms. This breaks the circular lock. When this RLock
fails, the job will be reported back to go-etcd-cron that the job is not
deliverable and unlock the queue lock- since the job _is_ actually
deliverable, go-etcd-cron will correctly re-enqueue the job immediately
whilst releasing that queue lock in the mean time. See:
https://github.com/diagridio/go-etcd-cron/blob/c9126ef694ffed83011125c5acbf6127f33bf784/internal/queue/queue.go#L209-L217

Fixes another deadlock in the Scheduler server whereby a connection to
daprd would not correctly be identified as a closed or cancelled
connection. This would happen when the connection is closed or
cancelled, whilst there is a send & job handle in progress for a
particular connection. This has been resolved by adding another go
routine to the connection handler to detect when either Recv or Send has
returned, the pool total is closed, or the stream context has closed,
then closing and removing that connection entirely. This correctly
ensures that the connection is removed from the pool, as well as
returning from the stream at the top gRPC RPC level.

Permanently backoff when calling a remote reminder when the remote peer
is not available, i.e. during a scale down event. This prevents future
retries, which slows down the entire placement dissemination for that
namespace.

Mark the placement client as ready later- after the placement client
has sent the first host request, so to better reflect the real ready
state and prevent unnecessary timing errors which cause retries
downstream.

Move `HaltAll` and `Drain` implementation to the same func as they were
the same implementation, only with one using a conditional for the
drain.

Pre-fetch the actor sub Interfaces on the Factory initialization of
workflow and activity actors. This prevents a circular deadlock during
cases of dissemination where a workflow/activity actor would be
deactivated whilst attempting to retrieve one of these interfaces (like
state).

Move the workflow actor internal lock to only be used in the _actual_
critical shared memory locations. Previous use of locking around actor
resource usage like state would cause the same circular deadlock as
described above.

The Scheudler client has been changed to be more aggressive for
scheduler reconnects, particularly on `WatchHosts`. This prevents
reconnect errors which were observed in scenarios whereby the entire
scheduler cluster was to go down temporarily. Also removes the retry
loop for the Scheduler client connector since a reconnect at the top
level is preferred.

Mark the Scheduler client readiness to true only when all Scheduler
host clients have actually sent the initial message to their respective
host. Sets readiness to false during shutdown and reconnect events.

JobsWatch passes a context up which will be closed when either
Send/Recv or shutdown is in progress to ensure the gRPC API is returned
on disconnect events.

Set Scheduler log level to `debug` in e2e CI runs.

Integration Tests:

Set all exec processes to use `debug` log level to make debugging easier
when Integration tests fail.

Fixes shutdown timeout of HTTP app Cleanup.
Fix log line process Cleanup logic to ensure all exec process logs are
flushed before checking for a log line existence.
Fixes timing errors in scheduler failure policy tests.
Adds actor type handler not implementing the deactivation endpoints.
Reduce the number of components in fuzz component name tests to prevent
shutdown timeouts on slow CI runnners.
Fix broken pipe network errors which occur on slow CI runners.
Adds tests for the new scheduler metadata field, both for gRPC and HTTP.
Fix state metrics tests by using the correct Enventually tester on
asserts.
Correctly use Workflow `WaitForSingleEvent` for workflow coordination,
rather than cross scope channels for singling.
Use eventuallys to prevent EOF errors during placement dissemination
tests.

Signed-off-by: joshvanl <me@joshvanl.dev>