[v1.15] Actors: Always attempt to connect to placement if configured.
Fixes a regression whereby daprd would no longer connect to placement if
an Actor state store is not configured. This meant an app which did not
have an actor state store configured (& host Actors), could not invoke
Actors on other apps.

Update daprd to always start Placement and the Actor runtime if
Placement addresses are configured.

Adds integration test for this scenario.

Adds metadatda integration tests.

Signed-off-by: joshvanl <me@joshvanl.dev>