.github/scripts/validate_sidecar_resources.py: Remove go routine check
Remove the go routine check from the stand-alone validation script. The
number of go routines is something we entirely don't care about-
scheduling of concurrent work onto threads is something that Go is
specifically designed to do, and does well. This check is not useful and
all it does it create red fatigue.

We have go routine leak checks in the long haul tests.

Signed-off-by: joshvanl <me@joshvanl.dev>