Prevent panic of reminder operations on slow Actor Startup
Prevent panic of reminder operations on slow Actor Startup

Problem

+The Dapr runtime HTTP server would panic if a reminder operation timed out while an Actor was starting up.

Impact

The HTTP server would panic, causing degraded performance.

Root cause

The Dapr runtime would attempt to use the reminder service before it was initialized.

Solution

Correctly return an errors that the actor runtime was not ready in time for the reminder operation.

Should be back-ported to release-1.15

Signed-off-by: joshvanl <me@joshvanl.dev>