Workflow: Rerun workflow instance from Event ID (#8699)
* Workflow: Rerun workflow instance form Event ID

Change extends the functionally with an rpc to enable rerunning a
workflow from an existing _terminal_ workflow instance. See;
https://github.com/dapr/proposals/pull/80

Rerunning workflows has the given restrictions;
- The source workflow instance _must_ be in a terminal state- i.e. is
  not running or suspended. It has to have been run before.
- The target event ID must be an activity. The event ID must exist.
- The target event ID input data can be overwritten.
- The new instance ID can be supplied, or will be randomly generated as
  a UUID.
- The source and target instance ID cannot be the same. The orginal
  history will not be overwritten/deleted/truncated.

The new workflow cannot have any timers or raise events which are active
at the time of starting. There is no technical resource why we cannot
achieve this, but is a limitation of the existing implementation. A
future change should see a refactor to enable this; likely moving timers
and raise events to a new actor type and instantiation.

To rerun a workflow, the source instance ID workflow actor will fork its
history state to the point of the target event ID, optionally update its
input, then send the new state to the target instance actor ID who will
then write that state and run the target activity, as well as  any
activity which are in-progress at that point in time. Care is taken to
ensure all currently active activities are executed if the target event
ID is in a fanout scenario.

```
rerunWorkfolow -> source Actor ID [Fork State] -{state}-> target Actor ID ->>> call activities
```

The workflow and activities actor code has been refactored to multiple
files for readability and maintainability.

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix actors timers callback timeout check

Signed-off-by: joshvanl <me@joshvanl.dev>

* Updates dapr/durabletask-go to v0.7.0

Signed-off-by: joshvanl <me@joshvanl.dev>

* Update dapr/durabletask-go to v0.7.1

Signed-off-by: joshvanl <me@joshvanl.dev>

* Adds review comments

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix parallel order execution timing

Signed-off-by: joshvanl <me@joshvanl.dev>

* Remove parallels

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix content length test

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>
Co-authored-by: Dapr Bot <56698301+dapr-bot@users.noreply.github.com>