[1.15] Update Max Concurrent Workflow Operations
Workflows would experience degraded performance in high throughput scenarios.

Workflows would be limited in the number of concurrent operations they could handle, leading to increased latency and potential timeouts.

The default value for `maxConcurrentWorkflowInvocations` and `maxConcurrentActivityInvocations` was set to 1000, which could lead to performance issues in high throughput scenarios.

Increase the default value for `maxConcurrentWorkflowInvocations` and `maxConcurrentActivityInvocations` to max int32 value, allowing for more concurrent operations to be processed without contention.

Signed-off-by: joshvanl <me@joshvanl.dev>