Scheduler: Allow for optional Etcd embed
Update the scheduler to allow for optional embedding of the Etcd
database. This is useful when running an external Etcd database.

When Etcd is not embedded (`--etcd-embed=false`), the scheduler
requires a `--etcd-client-endpoints` string array flag to be set,
which specifies the endpoints of the external Etcd cluster.

Scheduler continues to run Etcd embedded by default.

Signed-off-by: joshvanl <me@joshvanl.dev>