Workflow: Allow non-stream Workflow Schedule (#8763)
* [1.15] Workflow: Allow non-stream Workflow Schedule

Remove the registered actor type gate on the Scheduler workflow internal
server to allow Scheduling workflows on a daprd which does not have a
client listener stream.

This makes the Schedule Workflow RPC in line with the other Workflow
RPCs.

Problem

Attempting to schedule a Workflow on a daprd with no Workflow listener stream would hang.

Impact

Attempting to schedule a Workflow on a daprd with no Workflow listener stream would hang indefinitely.

Root cause

The daprd would wait until a Workflow listener stream was established before processing the schedule request.

Solution

Allow scheduling of Workflows without requiring a Workflow listener stream to be established first. This allows for immediate processing of the schedule request.

Signed-off-by: joshvanl <me@joshvanl.dev>

* Remove old register channel block

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>
Co-authored-by: Cassie Coyle <cassie@diagrid.io>
Co-authored-by: Dapr Bot <56698301+dapr-bot@users.noreply.github.com>