[1.15] Move Scheduler Actor Type Reload to Placement client
Moves the reloading of Sheduler Actor Types from the placement package
to the placement client. This fixes an issue with types being unreported
during failure scenarios connecting to placement.

Updates the placement client to be more aggressive on reconnecting and
on updating Actor Types.

Updates the Workflow Actor Backend to add a 15 second timeout to
operations to prevent a client side deadlock.

Signed-off-by: joshvanl <me@joshvanl.dev>