[1.15] Fixed Kafka PubSub intermittently dropping messages while retrying
Problem

Kafka PubSub would intermittently drop messages while retrying during rebalance or any other event requiring re-initialization of the consumer.

Impact

Messages sent to Kafka PubSub could be lost during retries, leading to potential data loss and inconsistencies in message processing.

Root cause

The processing loop for messages was continued in the consumer even after re-initialization, causing messages to no be processed if the consumer was not ready to process them.

Solution

Correctly breaking out of the processing loop in the Kafka consumer in the event of rebalancing or other re-initialization events.

Signed-off-by: joshvanl <me@joshvanl.dev>