[1.16] Prevent Jobs from not triggering upserting in high throughput scenarios (#8888)
* [1.16] Prevent Jobs from not triggering upserting in high throughput scenarios

Problem

Upserting a Job (or Actor reminder) in high job trigger throughput scenarios could result in the Job not being triggered.

Impact

Jobs (or Actor reminders) could not be triggered in high throughput scenarios.

Root cause

Scheduler would find a conflict in jobs being triggered between the old and new job state, and would result in no more jobs to be triggered on that host.

Solution

Ignore the conflict in the scheduler when upserting a job (or Actor reminder) in high throughput scenarios.

Signed-off-by: joshvanl <me@joshvanl.dev>

* Adds v1.15.7.md release notes

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>
Co-authored-by: Dapr Bot <56698301+dapr-bot@users.noreply.github.com>