[1.16] Integration: fix scheduler cluster wait for ready
Update the scheduler cluster framework "wait for ready" to check that
all of the schedulers have seen one another, and have settled
leadership. This prevents a race in scheduler cluster tests whereby
scheduler leaders are still trying to re-shuffle whilst the test has
already started. This manifests by "Job Already Exists" errors, as the
Job PUT call has it's context cancelled, but the key was successfully
written to the Etcd database disk.

Signed-off-by: joshvanl <me@joshvanl.dev>