Blocked Placement dissemination with high Scheduler dataset
Problem

Disseminations would hang for long periods of time when the Scheduler dataset was large.

Impact

Dissemination could take up to hours to complete, causing reminders to not be delivered for a long period of time.

Root Cause

The reminder migration of state store to scheduler reminders does a full decoded scan of the Scheduler database, which would take a long time if there were many entries. During this time the dissemination would be blocked.

Solution

Limit the maximum time spent doing the migration to 3 seconds.
Expose a new `global.reminders.skipMigration="true"` helm chart value which will skip the migration entirely.

Signed-off-by: joshvanl <me@joshvanl.dev>