Prevent infinite loop when workflow state is corrupted or destroyed (#9157)
* Prevent infinite loop when workflow state is corrupted or destroyed

Problem

Dapr workflows could enter an infinite reminder loop when the workflow state in the actor state store is corrupted or destroyed.

Impact

Dapr workflows would enter an infinite loop of reminder calls.

Root Cause

When a workflow reminder is triggered, the workflow state is loaded from the actor state store. If the state is corrupted or destroyed, the workflow would not be able to progress and would keep re-triggering the same reminder indefinitely.

Solution

Do not retry the reminder if the workflow state cannot be loaded, and instead log an error and exit the workflow execution.

Signed-off-by: joshvanl <me@joshvanl.dev>

* lint

Signed-off-by: joshvanl <me@joshvanl.dev>

* Add eventually for metrics

Signed-off-by: joshvanl <me@joshvanl.dev>

* Remove check which is not always correct

Signed-off-by: joshvanl <me@joshvanl.dev>

* Increase time for del state metrics check

Signed-off-by: joshvanl <me@joshvanl.dev>

* Wait for actor registration before continuing

Signed-off-by: joshvanl <me@joshvanl.dev>

* Adds sleep for actor registration

Signed-off-by: joshvanl <me@joshvanl.dev>

* Remove daprd3 from del state

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>
Co-authored-by: Dapr Bot <56698301+dapr-bot@users.noreply.github.com>