Scheduler: Don't install StatefulSet when Scheduler disabled (#9175)
* Scheduler: Don't install StatefulSet when Scheduler disabled

Now that the Dapr runtime client can resolve a single cluster IP service
DNS to multiple pod IPs, the scheduler no longer requires the full DNS
set of Scheduler replicas and Scheduler always runs with 3 replicas. We
therefore do not need the injector to discover the number of replicas of
Scheduler to determine if it is enabled or not.

Remove the StatefulSet resource completely when scheduler is disabled in
Kubernetes (`.Values.global.scheduler.enabled=false`).

Fixes https://github.com/dapr/dapr/issues/9138

Signed-off-by: joshvanl <me@joshvanl.dev>

* Fix unit test

Signed-off-by: joshvanl <me@joshvanl.dev>

* Update Dapr previous version

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>
Co-authored-by: Cassie Coyle <cassie@diagrid.io>