Workflow: schedule reminders to first history
Currently, all reminders work task reminders are sheduled with a due
time of `Now()`. The consequence of this is that processing a large
number of workflows concurrently leads to workflow execution favouring
breadth over depth.

PR updates the reminder scheduling logic to schedule reminders based on the
first history event of the workflow. This ensures that reminders are
scheduled in the order that workflows were created, leading to a more
depth first workflow execution pattern.

Signed-off-by: joshvanl <me@joshvanl.dev>