Workflow: schedule reminders to first history (#9179)
* Workflow: schedule reminders to first history

Currently, all reminders work task reminders are sheduled with a due
time of `Now()`. The consequence of this is that processing a large
number of workflows concurrently leads to workflow execution favouring
breadth over depth.

PR updates the reminder scheduling logic to schedule reminders based on the
first history event of the workflow. This ensures that reminders are
scheduled in the order that workflows were created, leading to a more
depth first workflow execution pattern.

Signed-off-by: joshvanl <me@joshvanl.dev>

* Pass through the start time of reminders to workflow activities on call

Signed-off-by: joshvanl <me@joshvanl.dev>

* Increase timeout of go routine test for placement

Signed-off-by: joshvanl <me@joshvanl.dev>

* Move activity reminder dueTime to request metadata

Signed-off-by: joshvanl <me@joshvanl.dev>

* Make due time metadata key a `const`

Signed-off-by: joshvanl <me@joshvanl.dev>

* Update log line with due time

Signed-off-by: joshvanl <me@joshvanl.dev>

* Spelling

Signed-off-by: joshvanl <me@joshvanl.dev>

---------

Signed-off-by: joshvanl <me@joshvanl.dev>
Co-authored-by: Cassie Coyle <cassie@diagrid.io>