Scheduler: worker pool
Workers is the number of workers that handle job events. The higher the number
the more go routines will be spawned, each working over a partition of the
total job space. The higher the number, the higher the number of jobs which
can be concurrently delivered to runtimes. Increasing this number increases
the number of go routines for this instance. This number should be tuned to
the bottleneck of job execution, i.e., CPU, I/O, memory, etc.

Defaults to 128 workers.

`--workers=N` flag for the scheduler binary.

`dapr_scheduler.workers=N` in the Helm chart.

Branched from https://github.com/dapr/dapr/pull/9205

Uses https://github.com/diagridio/go-etcd-cron/pull/103

Adds int test to ensure number of scheduler workers is configured

Signed-off-by: joshvanl <me@joshvanl.dev>