Deleted Jobs in all prefix matching deleted Namespaces
Problem

Deleting a namespace in Kubernetes will delete all the associated jobs in that namespace.
If there are any other namespaces with a name which has a prefix matching the deleted namespace, the jobs in those namespaces will also be deleted (i.e. deleting namespace "test" will also delete jobs in namespace "test-1" or "test-abc").

Impact

Deleting a namespace will delete jobs in other namespaces with prefix matching the deleted namespace.

Root Cause

Prefix logic did not terminate the prefix match with an exact match so that deleting a namespace would delete jobs in other namespaces with prefix matching the deleted namespace.

Solution

The prefix logic has been updated to ensure that only jobs in the exact deleted namespace are deleted.

Signed-off-by: joshvanl <me@joshvanl.dev>